+++ .eslintrc
@@ -26,6 +26,8 @@
     "jest/no-focused-tests": "error",
     "jest/no-identical-title": "error",
     "jest/prefer-to-have-length": "warn",
-    "jest/valid-expect": "error"
+    "jest/valid-expect": "error",
+    "import/no-unresolved": "off",
+    "import/no-cycle": "off"
   }
 }
